+++ src/events.js
@@ -1,19 +1,19 @@
 /* global $, not, camelCase, str2arr, normName, matches, isEmptyObject, isPlainObject */
 
-(function () {
-    if ( typeof window.CustomEvent === "function" ) return false;
-
-    function CustomEvent ( event, params ) {
-        params = params || { bubbles: false, cancelable: false, detail: null };
-        var evt = document.createEvent( 'CustomEvent' );
-        evt.initCustomEvent( event, params.bubbles, params.cancelable, params.detail );
-        return evt;
-    }
-
-    CustomEvent.prototype = window.Event.prototype;
-
-    window.CustomEvent = CustomEvent;
-})();
+// (function () {
+//     if ( typeof window.CustomEvent === "function" ) return false;
+//
+//     function CustomEvent ( event, params ) {
+//         params = params || { bubbles: false, cancelable: false, detail: null };
+//         var evt = document.createEvent( 'CustomEvent' );
+//         evt.initCustomEvent( event, params.bubbles, params.cancelable, params.detail );
+//         return evt;
+//     }
+//
+//     CustomEvent.prototype = window.Event.prototype;
+//
+//     window.CustomEvent = CustomEvent;
+// })();
 
 var overriddenStop =  Event.prototype.stopPropagation;
 var overriddenPrevent =  Event.prototype.preventDefault;